# docs/migration/contracts/api.yaml
...
openapi: 3.1.0
info:
  title: Silicon Extend API
  version: 2.0.0-draft
  description: |
    Silicon Extend lets a Silicon use a Carbon's personal devices. This API is used by the `extend` CLI
    (through the `silicon-extend-client` crate), the configuration website, and the Extend apps on devices.

    Conventions
    - Every JSON body is an envelope: `{"type": "<kind>", "data": {...}}`.
    - Errors: `{"type": "error", "data": {"code", "message", "hint", "docs_url", "request_id", "details"}}`.
      `code` is stable and machine-readable; `message` says what happened and why; `hint` says what to do next.
    - Timestamps are RFC 3339 UTC with milliseconds. Durations carry their unit in the field name.
    - Every response carries `X-Request-ID` and `Silicon-Extend-API-Version`.
    - Lists are paged with `limit` and `cursor`; the response gives `next_cursor`, null on the last page. A page that
      has a `next_cursor` is always full; only the last page can be short.
    - Retry a POST after an uncertain result with the same `Idempotency-Key` and the identical body. The same key with
      a different body is `409 conflict`.
    - API majors (TECHNICAL.md section 10). Every `/api/v{n}/` request is checked before anything else: a major this
      service doesn't serve is `400 api_version_unsupported`; a `Silicon-Extend-API-Version` pin naming another major
      than the path is `400 api_version_mismatch`; a sunset major is `410 api_version_sunset`, whose hint says how to
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v2/telemetry:
    post:
      tags: [Operations]
      operationId: telemetry
      summary: Forward a telemetry event to Space Station
      description: |
        So the CLI and apps never hold the Space Station ingest key. `X-Extend-Telemetry: off` drops it, and on
        any other route it keeps the service from recording that request (for example a command's event).
        Never accepts typed text, screen contents, tokens, codes or secrets.
      parameters:
        - {name: X-Extend-Telemetry, in: header, schema: {type: string, enum: ['on', 'off']}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, data]
              properties:
                type: {const: telemetry}
                data:
                  type: object
                  additionalProperties: false
                  required: [source, event, step, success, duration_ms]
                  properties:
                    source: {type: string, enum: [cli, client, app, web]}
                    event: {type: string, maxLength: 64, example: command}
                    step: {type: string, maxLength: 128, example: session.command.relay}
                    success: {type: boolean}
                    duration_ms: {type: integer, minimum: 0}
                    error_code: {type: [string, 'null'], maxLength: 64}
                    command: {type: [string, 'null'], maxLength: 64}
                    device_os: {oneOf: [{$ref: '#/components/schemas/DeviceOs'}, {type: 'null'}]}
                    session_id: {oneOf: [{$ref: '#/components/schemas/SessionIdValue'}, {type: 'null'}]}
                    request_id: {type: [string, 'null']}
                    client_version: {type: string, maxLength: 32}
      responses:
        '204': {description: Accepted or dropped because telemetry is off}

  # ───────────────────────────── Extend 4: accounts and custodians ─────────────────────────────
...
  /api/v2/silicons:
    get:
      tags: [Silicons]
      operationId: listSilicons
      summary: The Silicons a Carbon looks after, then the ones they gave access to
      description: |
        Carbons only. A Silicon a Carbon looks after (they are its custodian in Silicon Accounts) shows `looked_after:
        true` with how many grants and running sessions it has. A Silicon they gave access to shows how many of their
        devices it can use. Nobody else's Silicons are listed: there is no directory to browse.
      responses:
        '200':
          description: Silicons
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: silicons}
                  data:
                    type: object
                    required: [items]
                    properties:
                      items: {type: array, items: {$ref: '#/components/schemas/SiliconSummary'}}
        '403': {description: carbon_only, $ref: '#/components/responses/Error'}
  /api/v2/silicons/{silicon}:
    parameters:
      - {$ref: '#/components/parameters/SiliconPath'}
    get:
      tags: [Silicons]
      operationId: showSilicon
      summary: One Silicon the caller looks after or gave access to
      responses:
        '200':
          description: The Silicon
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: silicon}
                  data: {$ref: '#/components/schemas/SiliconSummary'}
        '403': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
  /api/v2/silicons/{silicon}/grants:
    parameters:
      - {$ref: '#/components/parameters/SiliconPath'}
    get:
      tags: [Silicons]
      operationId: siliconGrants
      summary: Every device a Silicon the caller looks after can use, whoever gave the access
      description: |
        The Silicon's custodian only (`403 no_access` otherwise). Each grant names the device, its OS and the Carbon
        who paired it. The custodian sees what its Silicon does in Extend and can stop it (renounce a grant, end a
        session, withdraw a wake request, keep a file); it never acts as the Silicon.
      responses:
        '200':
          description: Grants
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: access}
                  data:
                    type: object
                    required: [items]
                    properties:
                      items: {type: array, items: {$ref: '#/components/schemas/AccessGrant'}}
        '403': {$ref: '#/components/responses/Error'}
  /api/v2/silicons/{silicon}/grants/{device_id}:
    parameters:
      - {$ref: '#/components/parameters/SiliconPath'}
      - {$ref: '#/components/parameters/DeviceId'}
    delete:
      tags: [Silicons]
      operationId: renounceGrant
      summary: Give up a Silicon's access to a device (its custodian, or the Silicon itself)
      description: |
        Its running session on the device ends (`access_removed`) and its open wake requests there are withdrawn. The
        device's Carbon is told in the device's activity log.
      responses:
        '204': {description: Renounced}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}

  # ───────────────────────────── Silicon Accounts webhooks ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A Silicon Accounts access token for the app `extend`: an EdDSA-signed JWT whose `aud` is `extend` and whose
        `iss` is the Silicon Accounts this server trusts (`ACCOUNTS_URL`), verified with its published keys. Carbons get
        one with the Silicon Accounts device flow (the CLI) or the authorization-code flow (the website's server),
        Silicons by exchanging a short-lived token from `silicon-accounts login --app extend -q`. Tokens issued before
        a sign-out, removed access or deleted account are refused.
    enrollmentSecret:
      type: http
      scheme: bearer
      bearerFormat: 'Enrollment secret (ees_). Sent as "Authorization: Extend-Enrollment <secret>".'
    deviceCredential:
      type: http
      scheme: bearer
      bearerFormat: 'Device credential (edc_). Sent as "Authorization: Extend-Device <credential>".'
  parameters:
...
    DeviceId:
      name: device_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/DeviceIdValue'}
...
    SiliconPath:
      name: silicon
      in: path
      required: true
      description: A Silicon the caller looks after or gave access to, by its current `si:` id or its Silicon Accounts uuid
      schema: {type: string, example: 'si:chef'}
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            required: [type, data]
            properties:
              type: {const: error}
              data: {$ref: '#/components/schemas/Error'}
          example:
            type: error
            data:
              code: device_in_use
              message: 'Device 7c1e09ab (Saket''s Pixel) is being used by si:chef in session a3f since 2026-09-26T09:58:00Z. Only one Silicon can use a device at a time.'
              hint: 'Ask for it with: extend request send 7c1e09ab --reason "<why, up to 300 characters>"'
              docs_url: https://extend.teamofsilicons.com/docs/cli#error-device_in_use
              request_id: 01926f41-0a1b-7c2d-8e3f-4a5b6c7d8e9f
              details: {in_use: {silicon_id: 'si:chef', session_id: a3f, since: '2026-09-26T09:58:00.000Z'}}
          # When the Silicon using the device is on another side (another Carbon's pair, or outside the caller's custodian
          # circle), Extend names no one: "Device 7c1e09ab (Living room TV) is in use. Only one Silicon can use a device
          # at a time.", hint 'Ask for it: extend request send 7c1e09ab --reason "..."', details {in_use: {hidden: true}}.
...
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Stable. The full list is in cli.yaml under errors.
          example: device_in_use
        message: {type: string, description: 'What happened and why, in full'}
        hint: {type: [string, 'null'], description: 'What to do next, often an exact command'}
        docs_url: {type: [string, 'null'], format: uri}
        request_id: {type: string}
        details: {type: object, additionalProperties: true}
...
    DeviceIdValue: {type: string, pattern: '^[0-9a-f]{8}$', example: 7c1e09ab}
    SessionIdValue: {type: string, pattern: '^[0-9a-f]{3,}$', example: a3f, description: '3 hexadecimal characters, growing to 4 and more once every shorter id is used; never reused'}
...
    CarbonId: {type: string, pattern: '^c:', example: 'c:alice'}
    SiliconId: {type: string, pattern: '^si:', example: 'si:chef'}
    DeviceName: {type: string, minLength: 1, maxLength: 64, example: Saket's Pixel}
...
    DeviceOs:
      type: string
      enum: [android, android_tv, macos, windows, linux, ios, ipados, tvos, samsung_tv, lg_tv]
      description: android_tv covers Android TV, Google TV and Fire OS
...
    Member:
      type: object
      required: [type, id]
      properties:
        type: {type: string, enum: [carbon, silicon]}
        id: {type: string, example: 'c:alice', description: 'The current public id; it can change'}
        display_name: {type: [string, 'null']}
        uuid: {type: string, example: 750mDPTu, description: 'The permanent Silicon Accounts uuid (API v2; never on the device wire)'}
...
    SiliconSummary:
      type: object
      required: [uuid, id, looked_after, granted_by_you]
      properties:
        uuid: {type: string}
        id: {type: string, example: 'si:chef'}
        display_name: {type: string}
        pfp_url: {type: string, format: uri}
        looked_after: {type: boolean, description: 'The caller is its custodian: they see its grants, sessions, files and requests, and can stop them'}
        granted_by_you: {type: integer, minimum: 0, description: 'How many of the caller''s devices it can use'}
        grants: {type: integer, minimum: 0, description: 'Custodian views: every grant it holds, whoever gave it'}
        running_sessions: {type: integer, minimum: 0, description: 'Custodian views'}
...
    AccessGrant:
      type: object
      required: [device_id, silicon_id, granted_by, granted_at]
      properties:
        device_id: {$ref: '#/components/schemas/DeviceIdValue'}
        silicon_id: {$ref: '#/components/schemas/SiliconId'}
        granted_by: {$ref: '#/components/schemas/CarbonId'}
        granted_at: {type: string, format: date-time}
        last_used_at: {type: [string, 'null'], format: date-time}
        wake_muted: {type: boolean, description: '1.1: the owner turned this Silicon''s wake requests off on this pair'}
        silicon_uuid: {type: string, description: 'The Silicon''s permanent uuid (`silicon_id` is its current id)'}
        granted_by_uuid: {type: string, description: 'The granting Carbon''s permanent uuid'}
        device_name: {allOf: [{$ref: '#/components/schemas/DeviceName'}], description: 'Custodian views (GET /api/v2/silicons/{silicon}/grants)'}
        device_os: {allOf: [{$ref: '#/components/schemas/DeviceOs'}], description: Custodian views}
        owner: {allOf: [{$ref: '#/components/schemas/Member'}], description: 'Custodian views: the Carbon who paired the device'}
...